replace ImageJ viewer boolean bitwise-and with conditional-and
The ImageJ viewer tests the value of a bitwise-and expression on two
boolean values: "connected & !isConnected".  This is unusual.  It is
very likely the intention was to use a conditional-and: "connected &&
!isConnected".  Change this expression to use a conditional-and.

A conditional-and will short-circuit, whereas a bitwise-and will not.
In this particular case, the resulting value of the expressions will be
the same, but a conditional-and is the more appropriate operation.