HDF5 plugin support writing infinite number of frames
Setting NumCapture=0 now allows capturing any number of frames.

Note that this feature is not supported in combination with using the extra
virtual dimensions.

When using this feature, the HDF5 istorek parameter (size of chunk index
binary search tree) is calculated using an estimated sensible number of
frames.

This fixes issue #72