Replacing BOOST_REQUIRE_* macros with BOOST_CHECK* macros in unittests
This is because the REQUIRE bails out of the test case when failing - where as failing a CHECK
will allow the test case to continue running and giving more diagnostic information.

Use REQUIRE only when truly necessary (say in indexing and memory size operations).
Use CHECK by default or when in doubt.