NDArrayPool::report shows freeArray list details for details > 0
Changed the report method such that for details > 0 the details of
the arrays in the freelist are reported. This is most interesting
as memory status information as the freelist contains unused allocated
arrays that could be reused preventing reallocations. However if this
unused allocated memory is too large it may influence performance by
inducing paging operations.