Avoid double counting number of frames
The calculation of number and size of hdf5 cache slots is wrong when data
is 1D, in case of multiframe file, it shouldn't consider the 0 dimension,
as the number of frames is calculated separately.
In the case of non multiframe file, it will try to access the invalid
dimension -1.
The approach taken is checking that we are not in one  of those cases
before considering the dimension (this->rank-2), which would be 0 in
a multiframe file, and -1 in a non multiframe file