fix(NDPluginFile): possible memory leaks when using the Capture mode
Previously, when using the Capture mode, if you started a new capture
without saving the previously captured frames to a file, the 'pCapture'
array holding the frames wouldn't be free'd properly, causing a memory
leak. Another way of causing such leak was by changing the capture mode
to stream without saving the frames to a file.

With this commit, when resetting NDFileNumCaptured before starting a new
capture, we make sure to also empty the capture buffer if it is not
empty yet, so that no buffer is leaked.

Signed-off-by: flowln <thiago.ferreira@lnls.br>