Remove bad NDAttribute cases
NDAttribute was being compared with NX attributes.
Not only did the code do nothing after the the comparison (it simply
fell to default case), this also risks breaking the compilation after new
additions to NDAttrDataType_t.

For example: NX_CHAR = 4, which is the same of NDAttrInt32.
Another example: after trying to add types to NDAttrDataType_t I got:

../ADApp/pluginSrc/NDFileNexus.cpp: In member function 'void NDFileNexus::constTextToDataType(char*, int, void*)':
../ADApp/pluginSrc/NDFileNexus.cpp:726:5: error: duplicate case value
     case NDAttrUndefined:
     ^~~~
../ADApp/pluginSrc/NDFileNexus.cpp:681:5: note: previously used here
     case NX_UINT8:
     ^~~~