Document unique id and time stamp records
These records have been implemented for a long time [1], but aren't
documented for end-users. Place them in the record table in a dedicated
section, reflecting the structure used in the source database template.

Describe both time stamp second and nanosecond in more depth, in an
attempt to make it clearer how to interpret them.

[1] 8bcf59c5 (Added dimensions, uniqueId and TimeStamp, 2008-04-29)