Deduplicate indexed record entries in the HDF5 docs
There are multiple PVs that are instantiated up to 9 dimensions. Instead
of listing them exhaustively, document them once with an N variable as
done with NDArraySize parameter. This makes the documentation less
bloated and easier to add more such PVs in a concise and less
error-prone way.

Improving the way such collapsed table entries are documented is
out-of-the-scope of this patch, and it will be dealt with in future
changes.

While here, correct "posIndexDim" record names to use uppercase 'P'.