Document HDF5 dataset fill value record
This record is available since 7a0d0176 (Added fillValue parameter to
the HDF5 file writer., 2015-09-22), but it was never properly
documented.

Document it under the metadata section as it is basically metadata
regarding the default value the readers should read when no value is
directly defined. This has been documented based on the corresponding
explanation from the HDF5 library function [1] that uses this value,
that is, H5Pset_fill_value().

[1] https://support.hdfgroup.org/documentation/hdf5/latest/group___d_c_p_l.html