Standardize the way numbered record names are documented
Some records are named after a fixed number of dimensions or items, and
therefore can be collectively documented. To collapse its name while
keeping the information of where the instance number goes, we use square
brackets with the variable N. However, the way the ranges are specified
is inconsistent, and how the user should interpret the brackets
substitution is undocumented. Moreover, we use the variable N for this
context, but there are also PVs with references to such variable name as
well, making it even more confusing for a newcomer.

Change that by standardizing on using angle brackets to specify the
instance number, making consistently documented in the entry
description. To avoid confusion with variables names, use K instead of
N, and provide an example for each record of an expanded record name, in
order to make sure that readers will correctly drop the brackets
characters.

Because NDPluginGather substitutes its instance number based on the
environment variable, properly annotate it as such. In this case, the
range is user-defined, and it is already documented in the description.